+++ Widgets/Resource/Widgets.intentdefinition
