+++ Widgets/WidgetsBundle/WidgetsBundle.swift
@@ -11,7 +11,9 @@ import SwiftUI
 @main
 struct WidgetsBundle: WidgetBundle {
     var body: some Widget {
-        Widgets()
+        MyCardWidget()
+        OpenAppLockScreenWidget()
+        QRCodeWidget()
         WidgetsLiveActivity()
     }
 }
